python_backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.17.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
Pillow>=9.5.0
numpy>=1.24.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6